Fixed missing optional cases as well as creating a common trait for readMessage[?] usage.
